% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_word.R
\name{next_word}
\alias{next_word}
\title{Ngram Model to predict next word
We built and train three ngram to check what will be the next word, we check first with the last 3 words,
if nothing is found, the last two and so the last. If there is no match, the word the most used is returned.}
\usage{
next_word(str1)
}
\arguments{
\item{str1}{a sentence or word, just the maximum last three words will be in the process.}
}
\value{
the five word with the higher frequency following the precedents words.
}
\description{
Ngram Model to predict next word
We built and train three ngram to check what will be the next word, we check first with the last 3 words,
if nothing is found, the last two and so the last. If there is no match, the word the most used is returned.
}
\details{
This function is the last test for the data science coursera course of Hopkins university
}
\seealso{
\code{filter, tolower,str_split}
}
\author{
Aurelien Roussille
}
